Refacto ClusterLoading
